requirements(1).txt
# Core ML and Deep Learning
tensorflow>=2.13.0
keras>=2.13.1
numpy>=1.21.0
scikit-learn>=1.0.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Data Processing
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
requests>=2.25.0

# Optional: For GPU support (uncomment if using GPU)
# tensorflow-gpu>=2.13.0

# Optional: For cloud deployment
# boto3>=1.20.0  # AWS
# google-cloud-storage>=2.0.0  # Google Cloud
# azure-storage-blob>=12.0.0  # Azure

# Optional: For web deployment
# flask>=2.0.0
# fastapi>=0.70.0
# uvicorn>=0.15.0
# streamlit>=1.0.0

# Development tools
jupyter>=1.0.0
notebook>=6.4.0
